This commit message is for a new file called `scripts/commit_message`. The purpose of this file is to provide a script that generates a concise git commit message based on the user's input. The script takes in two arguments, `MODEL` and `PROMPT`, which are used to generate the commit message.
The approach taken by the script is to use a machine learning model to generate the commit message based on the input provided by the user. The model is trained on a dataset of commit messages and their corresponding descriptions, and it uses this training data to generate new commit messages that are similar in style and structure to the training data.

The script also includes some error handling to ensure that the response from the machine learning model is valid and not null. If the response is null or empty, the script will print an error message to the console.

Overall, this commit message is a way of documenting the approach taken by the `scripts/commit_message` script, which is designed to make it easier for developers to write informative and consistent commit messages.